Fixed more errors

+++ Celestin.API/Models/CelestinModels/CelestinWithoutDiscoveryDto.cs
@@ -6,7 +6,7 @@ namespace Celestin.API.Models.CelestinModels
     {
         public int Id { get; set; }
         public string Name { get; set; }
-        public decimal Mass { get; set; }
+        public double Mass { get; set; }
         public decimal EquatorialDiameter { get; set; }
         public int SurfaceTemperature { get; set; }
         public DateTime DiscoveryDate { get; set; }
